+++ json_serializable/lib/src/helper_core.dart
@@ -57,39 +57,45 @@ abstract class HelperCore {
 InvalidGenerationSourceError createInvalidGenerationError(
   String targetMember,
   FieldElement field,
-  UnsupportedTypeError e,
+  UnsupportedTypeError error,
 ) {
   var message = 'Could not generate `$targetMember` code for `${field.name}`';
+
   String todo;
-
-  if (e.type is TypeParameterType) {
+  if (error.type is TypeParameterType) {
     message = '$message because of type '
-        '`${e.type.getDisplayString(withNullability: false)}` (type parameter)';
+        '`${error.type.getDisplayString(withNullability: false)}` (type parameter)';
 
-    todo = r'''
-To support type paramaters (generic types) you can:
-1) Use `JsonConverter`
-  https://pub.dev/documentation/json_annotation/latest/json_annotation/JsonConverter-class.html
-2) Use `JsonKey` fields `fromJson` and `toJson`
-  https://pub.dev/documentation/json_annotation/latest/json_annotation/JsonKey/fromJson.html
-  https://pub.dev/documentation/json_annotation/latest/json_annotation/JsonKey/toJson.html
-3) Set `JsonSerializable.genericArgumentFactories` to `true`
+    todo = '''
+To support type parameters (generic types) you can:
+$converterOrKeyInstructions
+* Set `JsonSerializable.genericArgumentFactories` to `true`
   https://pub.dev/documentation/json_annotation/latest/json_annotation/JsonSerializable/genericArgumentFactories.html''';
-  } else if (field.type != e.type) {
-    message = '$message because of type `${typeToCode(e.type)}`';
+  } else if (field.type != error.type) {
+    message = '$message because of type `${typeToCode(error.type)}`';
+  } else {
+    todo = '''
+To support the type `${error.type.element.name}` you can:
+$converterOrKeyInstructions''';
   }
 
-  final messageItems = [
-    '$message.',
-    e.reason,
-    if (todo != null) todo,
-  ];
-
   return InvalidGenerationSourceError(
-    messageItems.join('\n'),
+    [
+      '$message.',
+      if (error.reason != null) error.reason,
+      if (todo != null) todo,
+    ].join('\n'),
     element: field,
   );
 }
+
+@visibleForTesting
+const converterOrKeyInstructions = r'''
+* Use `JsonConverter`
+  https://pub.dev/documentation/json_annotation/latest/json_annotation/JsonConverter-class.html
+* Use `JsonKey` fields `fromJson` and `toJson`
+  https://pub.dev/documentation/json_annotation/latest/json_annotation/JsonKey/fromJson.html
+  https://pub.dev/documentation/json_annotation/latest/json_annotation/JsonKey/toJson.html''';
 
 /// Returns a [String] representing the type arguments that exist on
 /// [element].