+++ json_serializable/test/src/core_subclass_type_input.dart
@@ -1,9 +1,10 @@
 part of '_json_serializable_test_input.dart';
 
 @ShouldThrow(
-  r'''
+  '''
 Could not generate `fromJson` code for `mapView`.
-None of the provided `TypeHelper` instances support the defined type.''',
+To support the type `MapView` you can:
+$converterOrKeyInstructions''',
   element: 'mapView',
 )
 @JsonSerializable(createToJson: false)
@@ -12,9 +13,10 @@ class UnsupportedMapField {
 }
 
 @ShouldThrow(
-  r'''
+  '''
 Could not generate `fromJson` code for `listView`.
-None of the provided `TypeHelper` instances support the defined type.''',
+To support the type `UnmodifiableListView` you can:
+$converterOrKeyInstructions''',
   element: 'listView',
 )
 @JsonSerializable(createToJson: false)
@@ -23,9 +25,10 @@ class UnsupportedListField {
 }
 
 @ShouldThrow(
-  r'''
+  '''
 Could not generate `fromJson` code for `customSet`.
-None of the provided `TypeHelper` instances support the defined type.''',
+To support the type `_CustomSet` you can:
+$converterOrKeyInstructions''',
   element: 'customSet',
 )
 @JsonSerializable(createToJson: false)
@@ -36,9 +39,10 @@ class UnsupportedSetField {
 abstract class _CustomSet implements Set {}
 
 @ShouldThrow(
-  r'''
+  '''
 Could not generate `fromJson` code for `customDuration`.
-None of the provided `TypeHelper` instances support the defined type.''',
+To support the type `_CustomDuration` you can:
+$converterOrKeyInstructions''',
   element: 'customDuration',
 )
 @JsonSerializable(createToJson: false)
@@ -49,9 +53,10 @@ class UnsupportedDurationField {
 abstract class _CustomDuration implements Duration {}
 
 @ShouldThrow(
-  r'''
+  '''
 Could not generate `fromJson` code for `customUri`.
-None of the provided `TypeHelper` instances support the defined type.''',
+To support the type `_CustomUri` you can:
+$converterOrKeyInstructions''',
   element: 'customUri',
 )
 @JsonSerializable(createToJson: false)
@@ -62,9 +67,10 @@ class UnsupportedUriField {
 abstract class _CustomUri implements Uri {}
 
 @ShouldThrow(
-  r'''
+  '''
 Could not generate `fromJson` code for `customDateTime`.
-None of the provided `TypeHelper` instances support the defined type.''',
+To support the type `_CustomDateTime` you can:
+$converterOrKeyInstructions''',
   element: 'customDateTime',
 )
 @JsonSerializable(createToJson: false)